backend/workspace-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-api-python-client==2.150.0
google-auth==2.35.0
google-auth-oauthlib==1.2.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
anthropic>=0.39.0
fpdf2>=2.8.0
